+++ packages/garbo/src/resources/realm.ts
@@ -11,7 +11,7 @@ import { globalOptions } from "../config";
 import { garboValue } from "../garboValue";
 import { HIGHLIGHT } from "../lib";
 import { acquire } from "../acquire";
-import { shouldYachtzee } from "../yachtzee/lib";
+import { shouldYachtzee } from "../tasks/yachtzee/lib";
 
 type VolcanoItem = { quantity: number; item: Item; choice: number };
